Move io logic to separate file
