+++ index.js
@@ -31,7 +31,7 @@ const sessionMiddleware = session({
 });
 
 // Active users
-const users = [];
+// const users = [];
 
 // Connect to MongoDB
 require('./config/database');
@@ -51,28 +51,8 @@ app.use(sessionMiddleware);
 app.use(passport.initialize());
 app.use(passport.session());
 
-const wrap = (middleware) => (socket, next) =>
-  middleware(socket.request, {}, next);
-
-io.use(wrap(sessionMiddleware));
-io.use(wrap(passport.initialize()));
-io.use(wrap(passport.session()));
-
-// io Middleware
-// io.use((socket, next) => {
-//   // Now we can access the passport body in sockets
-//   // userid = socket.request.session.passport.user;
-//   sessionMiddleware(socket.request, {}, next);
-// });
-
-io.use((socket, next) => {
-  // console.log(socket.request.user);
-  if (socket.request.user) {
-    next();
-  } else {
-    next(new Error('unauthorized'));
-  }
-});
+// io
+require('./io/io')(io, sessionMiddleware, passport);
 
 // Connect flash middleware
 app.use(flash());
@@ -87,66 +67,12 @@ app.use(function (req, res, next) {
   next();
 });
 
-// Socket Handling
-io.on('connection', async (socket) => {
-  // console.log(socket.request);
-  // console.log('User connected!');
-  // Look for new user
-
-  socket.on('join', () => {
-    // socket.broadcast.emit('message', {
-    //   username: 'Bot 🤖',
-    //   message: `${socket.request.user.username} has joined the chat 🤩`,
-    // });
-    // if (!users.includes(socket.request.user.username)) {
-    // }
-    users.push(socket.request.user.username);
-
-    // Send the username of the loggedin user when user joins
-    socket.emit('join', {
-      username: socket.request.user.username,
-    });
-  });
-
-  socket.on('message', async (msg) => {
-    // Get the user who sent the message
-    // Emit to all users accept the current user
-    // data = {
-    //   username: socket.request.user.username,
-    //   message: msg,
-    // };
-    try {
-      let newmessage = new Message({
-        text: msg,
-        by: socket.request.user._id,
-      });
-
-      await newmessage.save();
-
-      socket.broadcast.emit('message', {
-        username: socket.request.user.username,
-        message: msg,
-      });
-    } catch (err) {
-      console.log('Error: ' + err);
-    }
-  });
-
-  socket.on('disconnect', () => {
-    // io.emit('message', {
-    //   username: 'Bot 🤖',
-    //   message: `${socket.request.user.username} has left the chat 😢`,
-    // });
-    users.pop(socket.request.user.username);
-  });
-});
-
 // TODO: better way to get the active users
 // Gets active users
-app.get('/active', (req, res) => {
-  // Set removes duplicates from the array
-  res.json([...new Set(users)]);
-});
+// app.get('/active', (req, res) => {
+//   // Set removes duplicates from the array
+//   res.json([...new Set(users)]);
+// });
 
 app.use(routesHandler);
